+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.1.2
 -----------------
 - Maintain list preferences as GET params in form URL (save) in Schedule.php
+- Use basename() to extract file name from path in StudentsUsersInfo.fnc.php, Accounting/functions.inc.php, StudentAssignments.fnc.php & Student_Billing/functions.inc.php
 
 Changes in 11.1.1
 -----------------